move versioned branches rewrite to a plugin
